add readme and modify to lr to accept float

+++ ml/__main__.py
@@ -7,7 +7,7 @@ if __name__ == '__main__':
 
     parser.add_argument('--batch_size', metavar='batch_size', type=int, help='batch size')
     parser.add_argument('--epochs', metavar='epochs', type=int, help='num epochs')
-    parser.add_argument('--lr', metavar='lr', type=int, help='learning rate')
+    parser.add_argument('--lr', metavar='lr', type=float, help='learning rate')
 
     args = parser.parse_args()
